checkpoint isolate useResizeHandling
